- Added state logs - changed dataType float to int - changed iotype serial to spi

+++ lib/ui/utils/data_types.dart
@@ -1,1 +1,1 @@
-enum DataType {bool, string, map, float}
+enum DataType {bool, string, map, int}
